added docs for class and refactoring code
