+++ lib/domain/sight.dart
@@ -1,9 +1,20 @@
 class Sight {
+  /// Название объекта
   final String name;
+
+  /// Ширина
   final double lat;
+
+  /// Долгота
   final double lon;
+
+  /// uri
   final String url;
+
+  /// описание
   final String details;
+
+  /// тип достопремечательности
   final String type;
 
   Sight({
