+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui/screen/sight_details.dart';
+import 'package:places/mocks.dart';
+import 'package:places/ui/screen/sight_details_screen.dart';
 
 void main() {
   runApp(const App());
@@ -15,7 +16,7 @@ class App extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const SightDetails(),
+      home: SightDetailsScreen(sight: mocks.first),
     );
   }
 }
